fix: type predicates for layouts & refactor (#43)
* refactor: stick to utils naming convention

* fix: type predicates forlayouts, misc. imports
