+++ src/MetricSelect/SelectMetricAction.tsx
@@ -1,7 +1,7 @@
 import { SceneObjectState, SceneObjectBase, SceneComponentProps } from '@grafana/scenes';
 import { Button } from '@grafana/ui';
 
-import { MetricSelectedEvent } from '../../shared';
+import { MetricSelectedEvent } from '../shared';
 
 export interface SelectMetricActionState extends SceneObjectState {
   title: string;
