+++ src/MetricSelect/previewPanel.ts
@@ -2,8 +2,8 @@ import { PromQuery } from '@grafana/prometheus';
 import { SceneCSSGridItem, SceneQueryRunner, SceneVariableSet } from '@grafana/scenes';
 
 import { getAutoQueriesForMetric } from '../autoQuery/getAutoQueriesForMetric';
-import { getVariablesWithMetricConstant, MDP_METRIC_PREVIEW, trailDS } from '../../shared';
-import { getColorByIndex } from '../../utils';
+import { getVariablesWithMetricConstant, MDP_METRIC_PREVIEW, trailDS } from '../shared';
+import { getColorByIndex } from '../utils';
 
 import { AddToExplorationButton } from './AddToExplorationsButton';
 import { NativeHistogramBadge } from './NativeHistogramBadge';
