+++ src/otel/api.ts
@@ -2,7 +2,7 @@ import { RawTimeRange, Scope } from '@grafana/data';
 import { getPrometheusTime, isValidLegacyName } from '@grafana/prometheus';
 import { config, getBackendSrv } from '@grafana/runtime';
 
-import { callSuggestionsApi } from '../../utils';
+import { callSuggestionsApi } from '../utils';
 
 import { OtelResponse, LabelResponse, OtelTargetType } from './types';
 import { limitOtelMatchTerms, sortResources } from './util';
