+++ src/utils.test.ts
@@ -6,7 +6,7 @@ import { MetricDatasourceHelper } from './helpers/MetricDatasourceHelper';
 import { sortResources } from './otel/util';
 import { VAR_OTEL_AND_METRIC_FILTERS } from './shared';
 import { getDatasourceForNewTrail, limitAdhocProviders } from './utils';
-import { isAdHocFiltersVariable } from 'utils/variables';
+import { isAdHocFiltersVariable } from 'utils/utils.variables';
 
 jest.mock('./TrailStore/TrailStore', () => ({
   getTrailStore: jest.fn(),
